MS: refactored styled components
